requirements.txt
torch
torchvision
opencv-python
moviepy
Pillow
accelerate
diffusers
transformers
torchaudio